modtrezorconfig: extract pin & unlocking to storage.c
