+++ SConscript.unix
@@ -12,6 +12,7 @@ LIBS_MOD = []
 SOURCE_MOD += [
     'embed/extmod/modtrezorconfig/modtrezorconfig.c',
     'embed/extmod/modtrezorconfig/norcow.c',
+    'embed/extmod/modtrezorconfig/storage.c',
 ]
 
 # modtrezorcrypto
